+++ anchorboosting/models.py
@@ -174,7 +174,9 @@ class AnchorBooster:
             )
 
         if hasattr(X, "columns"):
-            feature_name = X.columns
+            feature_name = list(X.columns)
+        else:
+            feature_name = None
 
         if _POLARS_INSTALLED and isinstance(X, pl.DataFrame):
             X = X.to_arrow()